Fix issue with <= 0 fees that should be None in assert location
